Introduce new Hide Feedback setting
When the form is not meant to be known by non-group-members, this option
provides a way to completely hide any form output. When unchecked, the
plugin provides a 'Sorry. You are not allowed to view this form.' message
for users who are not a member of any of the form's groups.